+++ src/custom/state/orders/middleware.ts
@@ -6,17 +6,7 @@ import * as OrderActions from './actions'
 
 import { OrderIDWithPopup, OrderTxTypes, PopupPayload, buildCancellationPopupSummary, setPopupData } from './helpers'
 import { registerOnWindow } from 'utils/misc'
-
-type SoundType = 'SEND' | 'SUCCESS' | 'ERROR'
-type Sounds = Record<SoundType, string>
-
-const COW_SOUNDS: Sounds = {
-  SEND: '/audio/mooooo-send__lower-90.mp3',
-  SUCCESS: '/audio/mooooo-success__ben__lower-90.mp3',
-  ERROR: '/audio/mooooo-error__lower-90.mp3',
-}
-
-const SOUND_CACHE: Record<string, HTMLAudioElement | undefined> = {}
+import { getCowSoundError, getCowSoundSend, getCowSoundSuccess } from 'utils/sound'
 
 // action syntactic sugar
 const isSingleOrderChangeAction = isAnyOf(
@@ -171,30 +161,6 @@ export const popupMiddleware: Middleware<Record<string, unknown>, AppState> = (s
   })
 
   return result
-}
-
-function getAudio(type: SoundType): HTMLAudioElement {
-  const soundPath = COW_SOUNDS[type]
-  let sound = SOUND_CACHE[soundPath]
-
-  if (!sound) {
-    sound = new Audio(soundPath)
-    SOUND_CACHE[soundPath] = sound
-  }
-
-  return sound
-}
-
-function getCowSoundSend(): HTMLAudioElement {
-  return getAudio('SEND')
-}
-
-function getCowSoundSuccess(): HTMLAudioElement {
-  return getAudio('SUCCESS')
-}
-
-function getCowSoundError(): HTMLAudioElement {
-  return getAudio('ERROR')
 }
 
 function removeLightningEffect() {
